refactor(studio): import SQL editor store from source, delete facade + barrel
Migrate all 23 consumers of the @/state/sql-editor-v2 facade to import directly
from @/state/sql-editor/sql-editor-state (where useSqlEditorV2StateSnapshot,
getSqlEditorV2StateSnapshot, useSnippets and useSnippetFolders actually live).

Delete the sql-editor-v2.ts facade and the sql-editor/index.ts barrel — nothing
imports them anymore. This drops the two-layer re-export (v2 -> index -> source)
in favour of importing from the source layer, per the repo's no-barrel
convention.

Pure import-path migration; no behavior change. Symbol names keep their V2
suffix for now (renaming is a separate, larger churn).
