+++ apps/studio/components/interfaces/SQLEditor/UtilityPanel/SavingIndicator.tsx
@@ -5,7 +5,7 @@ import { Button, Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 
 import ReadOnlyBadge from './ReadOnlyBadge'
 import { useProfile } from '@/lib/profile'
-import { useSqlEditorV2StateSnapshot } from '@/state/sql-editor-v2'
+import { useSqlEditorV2StateSnapshot } from '@/state/sql-editor/sql-editor-state'
 import { isSaveFailed, isSaving } from '@/state/sql-editor/sql-editor-lifecycle'
 import { isSnippetOwner } from '@/state/sql-editor/sql-editor-rules'
 import { useSqlEditorSaveCoordinator } from '@/state/sql-editor/sql-editor-save-coordinator'
