--- apps/studio/state/sql-editor-v2.ts
@@ -1,8 +1,0 @@
-export type { StateSnippet, StateSnippetFolder } from './sql-editor'
-export {
-  sqlEditorState,
-  getSqlEditorV2StateSnapshot,
-  useSqlEditorV2StateSnapshot,
-  useSnippetFolders,
-  useSnippets,
-} from './sql-editor'
