--- apps/studio/state/sql-editor/index.ts
@@ -1,8 +1,0 @@
-export type { StateSnippet, StateSnippetFolder } from './types'
-export {
-  sqlEditorState,
-  getSqlEditorV2StateSnapshot,
-  useSqlEditorV2StateSnapshot,
-  useSnippetFolders,
-  useSnippets,
-} from './sql-editor-state'
